packages/paso-py/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "usepaso"
version = "0.3.1"
description = "Make your API agent-ready in minutes. One declaration, every protocol."
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
readme = "README.md"
keywords = ["ai", "agents", "mcp", "a2a", "api", "sdk", "agent-readiness", "openapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.27",
    "click>=8.0",
    "mcp>=1.9.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "mypy>=1.10",
    "types-PyYAML>=6.0"
]

[project.urls]
Homepage = "https://usepaso.dev"
Documentation = "https://github.com/5h1vmani/usepaso"
Repository = "https://github.com/5h1vmani/usepaso"
Issues = "https://github.com/5h1vmani/usepaso/issues"
Changelog = "https://github.com/5h1vmani/usepaso/releases"

[project.scripts]
usepaso = "usepaso.cli:main"
